feat(user): :sparkles: Add model transform, pre-save hook, static function exists()
Schema options have a property called toObject which transforms document to javascript object, this takes the final object retreived from mongo, so to prevent password and _id from beign retreived, tranform() function takes the object and mutates the properties to set it to undefined. Pre save hooks are usefull to hash a password, and other type of procedures, also other hooks are avilable instead of 'pre'. Static methods are interesting, to abstract some functions that are only applicable to that model, very usefull in some scenarios.